requirements.txt
Flask==3.1.0
twilio==9.3.6
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
openpyxl==3.1.5
gunicorn==23.0.0
